new async implementation for Ray
